HT: encode

+++ src/lib/jp2/transform/transform/WaveletForward.h
@@ -18,7 +18,6 @@
 #pragma once
 
 #include "grok_includes.h"
-#include <cmath>
 
 namespace grk {
 
@@ -91,11 +90,11 @@ template <typename DWT> bool WaveletForward<DWT>::run(TileComponent *tilec){
 
 		// transform vertical
 		if (rw) {
-			const uint32_t linesPerThreadV = static_cast<uint32_t>(std::ceil((float)rw / hardware_concurrency()));
+			const uint32_t linesPerThreadV = static_cast<uint32_t>(std::ceil((float)rw / (float)hardware_concurrency()));
 			const uint32_t s_n = rh_next;
 			const uint32_t d_n = rh - rh_next;
 			std::vector< std::future<int> > results;
-			for(size_t i = 0; i < hardware_concurrency(); ++i) {
+			for(uint32_t i = 0; i < hardware_concurrency(); ++i) {
 				uint32_t index = i;
 				results.emplace_back(
 					Scheduler::g_tp->enqueue([this, index, bj_array,a,
@@ -126,10 +125,10 @@ template <typename DWT> bool WaveletForward<DWT>::run(TileComponent *tilec){
 		if (rh){
 			const uint32_t s_n = rw_next;
 			const uint32_t d_n = rw - rw_next;
-			const uint32_t linesPerThreadH = static_cast<uint32_t>(std::ceil((float)rh / hardware_concurrency()));
+			const uint32_t linesPerThreadH = static_cast<uint32_t>(std::ceil((float)rh / (float)hardware_concurrency()));
 			std::vector< std::future<int> > results;
-			for(size_t i = 0; i < hardware_concurrency(); ++i) {
-				size_t index = i;
+			for(uint32_t i = 0; i < hardware_concurrency(); ++i) {
+				uint32_t index = i;
 				results.emplace_back(
 					Scheduler::g_tp->enqueue([this, index, bj_array,a,
 												 stride, rw,rh,
